Remove UI code from OneToOneCommunication file
We need to keep isolated the call and methods for the SDK for
tokbox, the class OneToOneCommunication is the only one how is
going to have this functionality.

We remove the .xib file and move the images to the Assets.xcassets
structure.

Also, we refactored the code from the controller to have the actions
for the UI.