Fix listeners being called many times
We have on the 'init' method the callback for connectionCreated
to call the Communication component (A/V component), so when a
user connects this component was called many times (increase according
with the times that the other party ends and start a call.

To fix it, we need to change the callback from connectionCreated to
sessionConnected which only involves one time call for the component.